+++ tests/PdfCompareLinkTest.php
@@ -14,6 +14,7 @@ declare(strict_types=1);
 namespace fpdf\Tests;
 
 use fpdf\PdfDocument;
+use fpdf\Tests\Fixture\FPDF;
 
 final class PdfCompareLinkTest extends AbstractCompareTestCase
 {
@@ -25,7 +26,7 @@ final class PdfCompareLinkTest extends AbstractCompareTestCase
     }
 
     #[\Override]
-    protected function updateOldDocument(\FPDF $doc): void
+    protected function updateOldDocument(FPDF $doc): void
     {
         $doc->Link(10, 10, 100, 5, 'https://wwww.bibi.nu');
         $doc->AddLink();