+++ tests/PdfCompareMarginsTest.php
@@ -14,6 +14,7 @@ declare(strict_types=1);
 namespace fpdf\Tests;
 
 use fpdf\PdfDocument;
+use fpdf\Tests\Fixture\FPDF;
 
 final class PdfCompareMarginsTest extends AbstractCompareTestCase
 {
@@ -28,7 +29,7 @@ final class PdfCompareMarginsTest extends AbstractCompareTestCase
     }
 
     #[\Override]
-    protected function updateOldDocument(\FPDF $doc): void
+    protected function updateOldDocument(FPDF $doc): void
     {
         $doc->SetLeftMargin(20);
         $doc->SetTopMargin(20);