+++ tests/PdfCompareMetadataTest.php
@@ -14,6 +14,7 @@ declare(strict_types=1);
 namespace fpdf\Tests;
 
 use fpdf\PdfDocument;
+use fpdf\Tests\Fixture\FPDF;
 
 final class PdfCompareMetadataTest extends AbstractCompareTestCase
 {
@@ -28,7 +29,7 @@ final class PdfCompareMetadataTest extends AbstractCompareTestCase
     }
 
     #[\Override]
-    protected function updateOldDocument(\FPDF $doc): void
+    protected function updateOldDocument(FPDF $doc): void
     {
         $doc->SetTitle('Title æ', true);
         $doc->SetSubject('Subject');